dxf: extract BaseExternalMeta and plan-meta paths into dxfutil
Move BaseExternalMeta, its external/internal field marshalers, and the
PlanMetaPath/PreparedMetaPath/SubtaskMetaPath helpers from
ingestor/globalsort into dxf/framework/dxfutil, so DXF task types (ddl
backfill, import-into, export) share them without depending on the
ingestion package.
