+++ pkg/dxf/framework/dxfutil/BUILD.bazel
@@ -2,12 +2,16 @@ load("@io_bazel_rules_go//go:def.bzl", "go_library", "go_test")
 
 go_library(
     name = "dxfutil",
-    srcs = ["util.go"],
+    srcs = [
+        "externalmeta.go",
+        "util.go",
+    ],
     importpath = "github.com/pingcap/tidb/pkg/dxf/framework/dxfutil",
     visibility = ["//visibility:public"],
     deps = [
         "//pkg/domain/sqlsvrapi",
         "//pkg/dxf/framework/storage",
+        "//pkg/objstore/storeapi",
         "//pkg/sessionctx",
         "@com_github_pingcap_errors//:errors",
     ],
@@ -16,13 +20,17 @@ go_library(
 go_test(
     name = "dxfutil_test",
     timeout = "short",
-    srcs = ["util_test.go"],
+    srcs = [
+        "externalmeta_test.go",
+        "util_test.go",
+    ],
     embed = [":dxfutil"],
     flaky = True,
-    shard_count = 2,
+    shard_count = 5,
     deps = [
         "//pkg/domain/sqlsvrapi/mock",
         "//pkg/kv",
+        "//pkg/objstore",
         "//pkg/sessionctx",
         "//pkg/util",
         "//pkg/util/mock",
