+++ pkg/dxf/importinto/BUILD.bazel
@@ -30,6 +30,7 @@ go_library(
         "//pkg/domain/serverinfo",
         "//pkg/dumpformat/parquetfile",
         "//pkg/dxf/framework/dxfmetric",
+        "//pkg/dxf/framework/dxfutil",
         "//pkg/dxf/framework/handle",
         "//pkg/dxf/framework/metering",
         "//pkg/dxf/framework/planner",
@@ -126,6 +127,7 @@ go_test(
         "//pkg/domain",
         "//pkg/domain/serverinfo",
         "//pkg/domain/sqlsvrapi/mock",
+        "//pkg/dxf/framework/dxfutil",
         "//pkg/dxf/framework/handle",
         "//pkg/dxf/framework/mock",
         "//pkg/dxf/framework/planner",
