+++ pkg/ingestor/globalsort/BUILD.bazel
@@ -64,7 +64,7 @@ go_test(
     ],
     embed = [":globalsort"],
     flaky = True,
-    shard_count = 41,
+    shard_count = 38,
     deps = [
         "//pkg/dxf/framework/taskexecutor/execute",
         "//pkg/ingestor/engineapi",
